Restructure repositories feature folder
